Add a non_retry_mode configurable on NonInvite
This is an addition to add a new mode to handle transaction that are non INVITE
This allow the user to set the no_retry_mode? and a specific timeout.
When the no_retry_mode is enabled we won't make any retry within the SIP Transaction. We will also send a timeout that can be set aswell in the config.

A use case for that is when we use the library to send SIP Options and we wnat to see if a server responds within X miliseconds